feat: 프록시 활용하여 예외 메시지 출력 기능 개발
